STARK: use GL_UNSIGNED_SHORT for glDrawElements instead of GL_UNSIGNED_INT (not supported for GLES <= 2)

+++ engines/stark/model/model.h
@@ -48,7 +48,7 @@ public:
 
 struct Face {
 	uint32 materialId;
-	Common::Array<uint32> vertexIndices;
+	Common::Array<uint16> vertexIndices;
 
 	Face() : materialId(0) {}
 };
